tests/test_aws_scanner.py
import pytest
from unittest.mock import Mock, patch
from modules.aws_scanner import AWSScanner

class TestAWSScanner:
    @pytest.fixture
    def mock_aws_clients(self):
        return {
            'ec2': Mock(),
            'eks': Mock(),
            'lambda': Mock(),
            'securityhub': Mock(),
            'inspector2': Mock()
        }
    
    def test_scan_ec2_instances(self, mock_aws_clients):
        scanner = AWSScanner(mock_aws_clients)
        # Test implementation
        pass
    
    # More test methods...