requirements.txt
streamlit>=1.28.0,<1.30.0
boto3>=1.28.0,<1.29.0
python-dotenv>=1.0.0,<1.1.0
pandas>=2.2.3,<3.0.0
plotly>=5.15.0,<6.0.0
openpyxl>=3.1.0,<3.2.0